@@ -30,18 +30,27 @@
     "@kyma-project/luigi-core": "0.4.11",
     "@types/async": "^2.0.49",
     "angular-sortablejs": "^2.6.0",
+    "apollo-angular": "^1.5.0",
+    "apollo-angular-link-http": "^1.6.0",
+    "apollo-cache-inmemory": "^1.5.1",
+    "apollo-client": "^2.5.1",
+    "apollo-link": "^1.2.11",
+    "apollo-utilities": "^1.2.1",
     "brace": "0.11.1",
     "core-js": "^2.5.7",
     "file-saver": "^1.3.8",
     "fiori-fundamentals": "1.4.3",
     "fundamental-ngx": "0.2.0",
-    "js-yaml": "^3.12.0",
+    "graphql": "^14.2.1",
+    "graphql-tag": "^2.10.1",
+    "js-yaml": "^3.13.1",
     "jsoneditor": "^5.24.3",
     "lodash": "^4.17.11",
     "ng-click-outside": "^4.0.0",
     "ngx-clipboard": "^11.1.7",
     "rxjs": "^6.3.1",
     "sortablejs": "^1.7.0",
+    "subscriptions-transport-ws": "^0.9.16",
     "time-ago-pipe": "^1.3.2",
     "zone.js": "^0.8.26"
   },
@@ -52,7 +61,7 @@
     "@angular-devkit/core": "0.6.8",
     "@angular-devkit/schematics": "^0.6.8",
     "@angular/cli": "^6.1.5",
-    "@angular/compiler-cli": "^6.1.6",
+    "@angular/compiler-cli": "^7.2.7",
     "@angular/language-service": "^6.1.6",
     "@babel/core": "^7.3.3",
     "@babel/preset-env": "^7.2.3",
@@ -73,12 +82,12 @@
     "karma-jasmine-html-reporter": "~1.4.0",
     "node-sass": "^4.11.0",
     "protractor": "^5.4.0",
-    "puppeteer": "1.11.0",
+    "puppeteer": "^1.14.0",
     "ts-node": "~3.2.0",
     "tslint": "^5.11.0",
     "tslint-angular": "^1.1.2",
     "tslint-config-prettier": "^1.15.0",
-    "typescript": "~2.9.2",
+    "typescript": "~3.1.1",
     "webpack": "^4.27.1",
     "webpack-cli": "^3.0.3",
     "webpack-dev-server": "^3.1.14"
