@@ -52,7 +52,9 @@ export class FilteredApisEntryRendererComponent
   }
 
   public ngOnDestroy() {
-    this.currentNamespaceSubscription.unsubscribe();
+    if (this.currentNamespaceSubscription) {
+      this.currentNamespaceSubscription.unsubscribe();
+    }
     this.communicationServiceSubscription.unsubscribe();
   }
 
