@@ -8,10 +8,10 @@ import { ComponentCommunicationService } from '../../../../../shared/services/co
 import { AppConfig } from '../../../../../app.config';
 import { Filter } from 'app/generic-list';
 import { Subscription } from 'rxjs';
-import { GraphQLClientService } from '../../../../../shared/services/graphql-client-service';
 import { GraphQLDataProvider } from '../../../operation/graphql-data-provider';
 import { ActivatedRoute } from '@angular/router';
 import { IEmptyListData } from 'shared/datamodel';
+import { GraphQLClientService } from 'shared/services/graphql-client-service';
 
 @Component({
   selector: 'app-filtered-apis',
@@ -73,7 +73,6 @@ export class FilteredApisComponent
         }namespaces/${namespaceId}/apis`;
 
         this.source = new GraphQLDataProvider(
-          `${AppConfig.graphqlApiUrl}`,
           query,
           {
             namespace: this.currentNamespaceId,
@@ -95,6 +94,8 @@ export class FilteredApisComponent
   }
 
   public ngOnDestroy() {
-    this.currentNamespaceSubscription.unsubscribe();
+    if (this.currentNamespaceSubscription) {
+      this.currentNamespaceSubscription.unsubscribe();
+    }
   }
 }
