@@ -8,11 +8,10 @@ import { AbstractKubernetesElementListComponent } from './abstract-kubernetes-el
 import { GraphqlMutatorModalComponent } from 'shared/components/json-editor-modal/graphql-mutator-modal.component';
 import { CurrentNamespaceService } from 'namespaces/services/current-namespace.service';
 import { ComponentCommunicationService } from 'shared/services/component-communication.service';
-import { GraphQLClientService } from 'shared/services/graphql-client-service';
 import { Filter } from 'app/generic-list';
 import { Subscription } from 'rxjs';
 import { GraphQLDataProvider } from './graphql-data-provider';
-import { AppConfig } from 'app/app.config';
+import { GraphQLClientService } from 'shared/services/graphql-client-service';
 
 @Component({
   selector: 'abstract-graphql-element-list',
@@ -39,37 +38,44 @@ export class AbstractGraphqlElementListComponent
     changeDetector: ChangeDetectorRef
   ) {
     super(currentNamespaceService, changeDetector, null, commService);
-
     this.currentNamespaceSubscription = this.currentNamespaceService
       .getCurrentNamespaceId()
       .subscribe(namespaceId => {
         this.currentNamespaceId = namespaceId;
         this.source = new GraphQLDataProvider(
-          AppConfig.graphqlApiUrl,
           this.getGraphqlQueryForList(),
           {
             namespace: this.currentNamespaceId
           },
-          this.graphQLClientService
+          this.graphQLClientService,
+          this.getGraphqlSubscriptionsForList(),
+          this.resourceKind
         );
       });
   }
 
   public ngOnDestroy() {
-    this.currentNamespaceSubscription.unsubscribe();
+    if (this.currentNamespaceSubscription) {
+      this.currentNamespaceSubscription.unsubscribe();
+    }
   }
 
   protected getGraphqlQueryForList() {
     return null; // override this
   }
 
+  protected getGraphqlSubscriptionsForList() {
+    return null; // override this
+  }
+
   editEntryEventCallback(entry) {
     const query = this.getResourceJSONQuery();
+    const variables = {
+      name: entry.name,
+      namespace: this.currentNamespaceId
+    };
     this.graphQLClientService
-      .request(AppConfig.graphqlApiUrl, query, {
-        name: entry.name,
-        namespace: this.currentNamespaceId
-      })
+      .gqlQuery(query, variables)
       .subscribe(data => {
         const lowerCaseResourceKind = this.resourceKind.charAt(0).toLowerCase() + this.resourceKind.slice(1);
         this.mutateResourceModal.resourceData = data[lowerCaseResourceKind].json;
@@ -79,30 +85,29 @@ export class AbstractGraphqlElementListComponent
 
   getResourceJSONQuery() {
     const lowerCaseResourceKind = this.resourceKind.charAt(0).toLowerCase() + this.resourceKind.slice(1);
-    return `query ${lowerCaseResourceKind}($name: String! $namespace: String!) {
-      ${lowerCaseResourceKind}(name: $name, namespace: $namespace) {
+    const variablesDefinitionsString = this.currentNamespaceId ? `$name: String!, $namespace: String!` :`$name: String!`;
+    const variablesString = this.currentNamespaceId ? `name: $name, namespace: $namespace` : `name: $name`;
+    return `query ${lowerCaseResourceKind}(${variablesDefinitionsString}) {
+      ${lowerCaseResourceKind}(${variablesString}) {
         json
       }
     }`;
   }
 
   sendDeleteRequest(entry) {
-    const name = entry.name;
-    const namespace = this.currentNamespaceId;
     const mutation = this.getDeleteMutation();
-
-    return this.graphQLClientService.request(
-      AppConfig.graphqlApiUrl,
-      mutation,
-      { name, namespace }
-    );
+    const variables = {
+      name: entry.name,
+      namespace: this.currentNamespaceId
+    };
+    return this.graphQLClientService.gqlMutation(mutation, variables);
   }
 
   getDeleteMutation() {
-    return `mutation delete${
-      this.resourceKind
-    }($name: String!, $namespace: String!) {
-      delete${this.resourceKind}(name: $name, namespace: $namespace) {
+    const variablesDefinitionsString = this.currentNamespaceId ? `$name: String!, $namespace: String!` :`$name: String!`;
+    const variablesString = this.currentNamespaceId ? `name: $name, namespace: $namespace` : `name: $name`;
+    return `mutation delete${this.resourceKind}(${variablesDefinitionsString}) {
+      delete${this.resourceKind}(${variablesString}) {
         name
       }
     }`;
