@@ -55,7 +55,9 @@ export class DeploymentEntryRendererComponent
 
   ngOnDestroy() {
     this.communicationServiceSubscription.unsubscribe();
-    this.currentNamespaceSubscription.unsubscribe();
+    if (this.currentNamespaceSubscription) {
+      this.currentNamespaceSubscription.unsubscribe();
+    }
   }
 
   isStatusOk(entry): boolean {
