@@ -191,7 +191,9 @@ export class ExposeApiComponent implements OnInit, OnDestroy {
   }
 
   public ngOnDestroy() {
-    this.currentNamespaceSubscription.unsubscribe();
+    if (this.currentNamespaceSubscription) {
+      this.currentNamespaceSubscription.unsubscribe();
+    }
   }
 
   private splitHostname(host) {
