@@ -66,7 +66,9 @@ export class ServiceDetailsComponent implements OnInit, OnDestroy {
   }
 
   public ngOnDestroy() {
-    this.currentNamespaceSubscription.unsubscribe();
+    if (this.currentNamespaceSubscription) {
+      this.currentNamespaceSubscription.unsubscribe();
+    }
   }
 
   public subscribeToRefreshComponent() {
